requirements.txt
pandas==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pyyaml~=6.0.1
openai==1.34.0
nltk==3.8.1
bs4~=0.0.2
numpy~=1.26.4
urllib3~=1.26.19